Split vector_tools.h
