+++ include/deal.II/numerics/vector_tools_constraints.templates.h
@@ -21,7 +21,7 @@
 
 #include <deal.II/hp/fe_values.h>
 
-#include <deal.II/numerics/vector_tools.h>
+#include <deal.II/numerics/vector_tools_constraints.h>
 
 DEAL_II_NAMESPACE_OPEN
